+++ src/web.rs
@@ -1,37 +1,17 @@
-use std::cmp::min;
 use std::path::Path;
-use futures::FutureExt;
-use futures::future::BoxFuture;
-use reqwest::{Client, Response};
+use follow_redirects::ClientExt;
+use hyper::Client;
+use hyper::body::HttpBody;
+use hyper::header::CONTENT_LENGTH;
+use hyper_tls::HttpsConnector;
 use tokio::fs::{File, self};
-use tokio::io::AsyncWriteExt;
-use futures::stream::StreamExt;
+use tokio::io::{AsyncWriteExt, BufWriter};
 use crate::config::Configuration;
 use crate::util::progress_bar::ProgressBar;
 
-pub fn request(url: &str) -> BoxFuture<'_, (Response, i32)> {
-    async move {
-        let response = Client::new()
-        .get(url)
-        .send()
-        .await
-        .unwrap();
-    
-    if let Some(size) = response.content_length() {
-        (response, size as i32)
-    } else {
-        request(url).await
-     }
-    }.boxed()
-}
-
 /// Downloads the jar from the configuration URL
-pub async fn download_server(config: &Configuration, target: &str) -> Result<(), std::io::Error> {
+pub async fn download_server(config: &Configuration, target: &str) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
     let server = &config.server;
-
-    let (response, max_size) = request(server).await;
-
-    let mut stream = response.bytes_stream();
 
     let temp_dir = Path::new(".temp");
     if !temp_dir.exists() {
@@ -39,26 +19,33 @@ pub async fn download_server(config: &Configuration, target: &str) -> Result<(),
     }
 
     let target_file = Path::new(target);
-
     let path_buf = temp_dir.join(target_file.file_name().unwrap());
-
     let temp_target_file = Path::new(path_buf.as_os_str());
 
-    let mut file = File::create(temp_target_file).await?;
+    let mut buffer = BufWriter::new(File::create(temp_target_file).await?);
 
-    let mut downloaded = 0;
+    let https = HttpsConnector::new();
+    let client = Client::builder()
+        .build::<_, hyper::Body>(https);
+    
+    let clip = server.parse()?;
 
-    let mut bar = ProgressBar::new(max_size);
+    let mut response = client.follow_redirects().get(clip).await?;
+    let total = response.headers()[CONTENT_LENGTH].to_str().unwrap().parse::<f32>()?;
 
-    while let Some(item) = stream.next().await {
-        let chunk = item.unwrap();
-        let _ = file.write(&chunk).await?;
-        downloaded = min(downloaded + chunk.len() as i32, max_size);
-        bar.set(downloaded);
+    let mut bar = ProgressBar::new(total as i32);
+    let mut read = 0;
+
+    while let Some(chunk) = response.body_mut().data().await {
+        let chunk = &chunk?;
+        read += chunk.len();
+        buffer.write_all(chunk).await?;
+        bar.set(read as i32);
         bar.print();
     }
 
-    let _ = fs::copy(temp_target_file, target_file).await.unwrap();
+    buffer.flush().await?;
+    let _ = fs::copy(temp_target_file, target_file).await?;
 
     // Exit Carriage Return
     bar.clear_text();
@@ -66,21 +53,22 @@ pub async fn download_server(config: &Configuration, target: &str) -> Result<(),
 }
 
 /// Downloads the file from the url and saves it to the target
-pub async fn download(url: String, target: String) -> Result<(), std::io::Error> {
-    let response = Client::new()
-        .get(&url)
-        .send()
-        .await
-        .expect("Error");
+pub async fn download(url: String, target: String) -> Result<(), Box<dyn std::error::Error + Send + Sync>> {
+    let mut buffer = BufWriter::new(File::create(&target).await?);
 
-    let mut file = File::create(&target).await?;
+    let https = HttpsConnector::new();
+    let client = Client::builder()
+        .build::<_, hyper::Body>(https);
+    
+    let uri = url.parse()?;
 
-    let mut stream = response.bytes_stream();
+    let mut response = client.follow_redirects().get(uri).await?;
 
-    while let Some(item) = stream.next().await {
-        let chunk = item.or(Err(format!("Error while downloading file"))).expect("Error");
-        file.write(&chunk).await?;
+    while let Some(chunk) = response.body_mut().data().await {
+        buffer.write_all(&chunk?).await?;
     }
+
+    buffer.flush().await?;
 
     Ok(())
 }